+++ NEWS.md
@@ -1,5 +1,6 @@
 # available (development version)
 
+* `get_wikipidia()` renamed to `get_wikipedia()`.
 * `valid_package_name()` now correctly checks a package name according to CRAN policy (@KevCaz, #61).
 
 # available 1.0.4